+++ src/client_src/gui/gui_animaciones.cpp
@@ -8,7 +8,7 @@ Animacion::Animacion() {}
 void Animacion::setFlip(bool flip) { this->flip = flip; }
 
 
-void Animacion::run(int pos_x, int pos_y, int speed, std::shared_ptr<std::vector<Frame>> frames,
+void Animacion::run(float pos_x, float pos_y, int speed, std::shared_ptr<std::vector<Frame>> frames,
                     std::vector<Frame>::iterator& it) {
     if (it == frames->end()) {
         it = frames->begin();
